Add mock gradingAPI
