+++ src/sagas/index.ts
@@ -7,7 +7,7 @@ import * as actions from '../actions'
 import * as actionTypes from '../actions/actionTypes'
 import { WorkspaceLocation } from '../actions/workspaces'
 import { mockAssessmentOverviews, mockAssessments } from '../mocks/assessmentAPI'
-import { mockGradingOverviews } from '../mocks/userAPI'
+import { mockGradingOverviews } from '../mocks/gradingAPI'
 import { IState } from '../reducers/states'
 import { Context, interrupt, runInContext } from '../slang'
 import { IVLE_KEY } from '../utils/constants'
